requirements.txt
flask
openai
PyPDF2
python-docx 